If jenkins is not able to attach a ssh retry up to 5 times

+++ jenkins.sh
@@ -7,6 +7,22 @@ if [ ! "$VBGUEST" ]; then
 fi;
 
 /opt/vagrant/bin/vagrant up
+n=0
+until [ $n -ge 5 ]
+do
+  /opt/vagrant/bin/vagrant ssh -c "echo ok"
+  RESULT=$?
+  if [ "$RESULT" -eq "0" ]
+  then
+    break
+  fi
+  n=$[$n+1]
+  sleep 5
+done
+if [ "$RESULT" -ne "0" ]
+then
+  exit $RESULT
+fi
 /opt/vagrant/bin/vagrant ssh -c 'cd /home/vagrant/dxr && make clean && make test'
 RESULT=$?
 /opt/vagrant/bin/vagrant destroy --force || exit $?
